+++ src/config/index.ts
@@ -73,10 +73,6 @@ const CONFIG = {
     ALQ_SAVINGS_ACCOUNT_ID: Number(process.env.ALQ_SAVINGS_ACCOUNT_ID),
     ALQ_API_KEY: process.env.ALQ_API_KEY,
   },
-  RECEIVING_BANK: {
-    BANK_ID: process.env.BANK_ID,
-    CLABE: process.env.CLABE,
-  },
   INTERNAL: {
     BOT_API_PORT: process.env.BOT_API_PORT,
     BOT_API_BASE_URL: process.env.BOT_API_BASE_URL,