+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [1.0.0] - 2024-07-17
 
 ### Added
-- Re-added ability to mark payments as paid by agent (meaning payment was made manually)
+- Re-added ability to mark payments as paid by agent (meaning payment was made manually).
+
+### Changed
+- Replaced `bankName` with `bankId` on `UserRoot.bankAccount`.
 
 ## [1.0.0] - 2024-07-16